NeuroScribe-Quantum-Package/requirements.txt
streamlit==1.46.0
python-dotenv==1.0.0
PyMuPDF==1.26.1
openai==1.3.7
Pillow==11.1.0
pytesseract==0.3.10
requests
pytest
bandit
safety
opencv-python==4.8.1.78
rembg==2.0.50
numpy==1.24.3
scikit-image==0.21.0
termcolor
pdfplumber==0.11.7
pdf2image==1.17.0
reportlab==4.4.2
